+++ src/DropDownHandling/AssignmentCheckbox.java
@@ -1,4 +1,4 @@
-package HaldingDropDown;
+package DropDownHandling;
 
 import java.util.List;
